+++ websift/__init__.py
@@ -24,7 +24,7 @@ Usage:
     )
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 from websift.client import WebSearchClient
 from websift.settings import AppSettings